handle body parsing in produced api functions
