+++ packages/gatsby/package.json
@@ -52,6 +52,7 @@
     "babel-preset-gatsby": "^3.12.0-next.1",
     "better-opn": "^2.1.1",
     "bluebird": "^3.7.2",
+    "body-parser": "1.20.1",
     "browserslist": "^4.21.7",
     "cache-manager": "^2.11.1",
     "chalk": "^4.1.2",
