+++ packages/gatsby/src/internal-plugins/functions/gatsby-node.ts
@@ -293,11 +293,9 @@ const createWebpackConfig = async ({
     )
 
     let entryToTheFunction = functionObj.originalAbsoluteFilePath
-    // if function has dynamic path, we inject it with webpack loader via query param
-    // see match-path-webpack-loader.ts for more info
-    if (functionObj.matchPath) {
-      entryToTheFunction += `?matchPath=` + functionObj.matchPath
-    }
+    // we wrap user defined function with our preamble that handles matchPath as well as body parsing
+    // see api-function-webpack-loader.ts for more info
+    entryToTheFunction += `?matchPath=` + (functionObj.matchPath ?? ``)
     entries[compiledNameWithoutExtension] = entryToTheFunction
   })
 
@@ -349,7 +347,7 @@ const createWebpackConfig = async ({
           test: /\.[tj]sx?$/,
           resourceQuery: /matchPath/,
           use: {
-            loader: require.resolve(`./match-path-webpack-loader`),
+            loader: require.resolve(`./api-function-webpack-loader`),
           },
         },
         {
